Fix shell option group handling in pkgconfig files
Since #1707 and #1710, `-Xarch_` option groups are present as single
elements in the libraries' COPTS, and a `SHELL:` prefix is added to
each of them. While these addressed certain build issues, they broke the
handling of `-Xarch_` option groups in pkgconfig files.

Fix that by taking care of the `SHELL:` prefix in COPTS when generating
pkgconfig files. The skip-next-flag mechanism is also removed as the
option groups are now present as single elements.

Signed-off-by: Ruoyu Zhong <zhongruoyu@outlook.com>