Remove `CharIterator::operator->`.
An `operator->` that doesn't return a record type doesn't really serve a useful
purpose, and it is not required by the iterator concepts. I haven't been able
to find any usage in non-test code, so I'm removing it.

Even though the `pointer` type alias is now not being used anywhere, I'm keeping
it because otherwise some standard libraries don't accept `CharIterator` when
passed to `std::string::apend()`, for example.

PiperOrigin-RevId: 659832167
Change-Id: I5de183a01009bfd71c8477f2023973075be8d860