Respect the allocator's .destroy method in ~InlinedVector
InlinedVector goes out of its way to use `A::construct` to construct new
elements, so it should not skip the `A::destroy` call for those elements,
either. Most codepaths seem fine (I didn't exhaustively explore this),
but `DestroyAdapter` specifically failed to check whether the allocator
had a non-trivial `destroy` method: it checked only whether the element
type was trivially destructible in the absence of any `destroy` method.

Use the same condition in `DestroyAdapter` as we already use in the
special short-circuit case on line 357 and in the conditions for
the relocation optimizations on lines 304 and 324.