Test that using a table that was moved-to from a moved-from table fails in sanitizer mode.
We previously added (and then rolled back) validation that caused moving a moved-from table to fail, but since we already validate the use-after-moved-to-from-moved-from case, it seems unnecessary to validate that we never move a moved-from table. Therefore, I don't think we need to pursue rolling forward the validation on moving a moved-from table.

PiperOrigin-RevId: 710749538
Change-Id: I64bbf7fd22fa160b47ad94fb8d51186e2a5bad3c