Minimize number of `InitializeSlots` with respect to SizeOfSlot.
I am still keeping `SooSlotMemcpySize` that may be 1, 4, 8 or 16. Microbenchmarks showed that it is useful to have compile time memcpy. In the long run, when/if more types will become trivially transferable and a potential increase of the SOO maximal size, this optimization can become (even) more useful. So I would stick with it for the time being.

`SooSlotMemcpySize` is always zero for non SOO types or not memcpyable types.
So as a result we merge all sizes for non SOO or non transferable types.
For memcpyable SOO types we have 4 options.

PiperOrigin-RevId: 723233539
Change-Id: If38c69527a2f873c07ce869bf3ca609fc7371a07