C++17 improvement: use if constexpr in internal/hash.h
As the minimal supported C++ version of Abseil is now C++17, it has become possible to use if constexpr statements. Apply usage of if constexpr in the header file hash.h.

Benefits of using if constexpr are:
- Improved readability: explicit that the expression will be evaluated at compile time.
- Required code is generated during compile phase instead of selected in optimizer phase. Improved guarantee that if statement is not checked at runtime.
- Prevents MSVC warning C4127: conditional expression is constant (when this warning is explicitly enabled when building Abseil with MSVC).

Remark: there are other header files that could be updated for if constexpr. This commit only addresses hash.h as recommended by the contributing  guidelines to keep PRs small.