Remove `WORKSPACE.bazel`
`MODULE.bazel` is now the recommended way to use Bazel.
Users that have wanted to use `WORKSPACE.bazel` still needed
to provide their own `WORKSPACE.bazel` file, so this file
provides no benefit to them. It was merely for us to test,
but there is little reason to do so.

Support for `WORKSPACE.bazel` is going away in Bazel 9.

PiperOrigin-RevId: 726117464
Change-Id: If6a2789eae2d236c0ec1b53162c7122c96967959