Add a log severity alias `DO_NOT_$UBMIT` intended for logging during development
'S' should be replace with S, in DO_NOT_$UBMIT :-)

`DO_NOT_$UBMIT` is an alias for `ERROR`, and is the right level to use
when you are using `LOG` for what's often called `printf()` debugging.
The name is easy to spot in review and is caught by common (but not
on-by-default) presubmit checks.

The contract is that **it must not be checked in**, so the Abseil team
reserves the right to delete it, change what it does, and/or scrub any
instances that do end up checked in, with or without notice.

Git users could consider adding a pre-commit hook to reject
the string "DO_NOT_$UBMIT" to use this feature.
https://git-scm.com/book/en/v2/Customizing-Git-Git-Hooks

PiperOrigin-RevId: 738364247
Change-Id: I3a277a941e7735ebf64b8da641bd42ca4830df9c