hold original pointer along with aligned to get rid of possibly lost memory leak
Unless pointers have been hold by an object, it's a lost on valgrind's logic. Even we don't free it, we must hold it. In this case, original pointers were not held.

Because gRPC uses this library, grpc usage was creating this memory leak.

before this change:
LEAK SUMMARY:
   definitely lost: 0 bytes in 0 blocks
   indirectly lost: 0 bytes in 0 blocks
     possibly lost: 1,408 bytes in 4 blocks
   still reachable: 1,408 bytes in 4 blocks
        suppressed: 0 bytes in 0 blocks

after this change:
LEAK SUMMARY:
   definitely lost: 0 bytes in 0 blocks
   indirectly lost: 0 bytes in 0 blocks
     possibly lost: 0 bytes in 0 blocks
   still reachable: 2,816 bytes in 8 blocks
        suppressed: 0 bytes in 0 blocks

Still reachable is not a problem for valgrind.

Signed-off-by: Oguzhan Turk <stkyoht@hotmail.com>