Expect different representations of pointers from the Windows toolchain.
In `TestPointerPrintedAsNumberDespiteAbslStringify`, an `EXPECT_DEATH` matches an error with a pointer represented in it. The representation used to differ for Windows. In newer releases, Windows' representation matches that of other toolchains.

This change makes either of the numeric representations admissible.

This change allows this test to pass before and after the latest Windows release.

PiperOrigin-RevId: 754003964
Change-Id: Ifee2c20a7d8022b380b7a58f50966013f7f7a4ba