Remove unnecessary modification of growth info in small table case.
We still allocate and allow modifying GrowthInfo in some codepathes. In most of the hottest codepathes we know that we don't need to modify or read growth_info for small tables.

It is still allowed to use `ResetGrowthLeft` and get access via `GetGrowthInfoFromControl`. That simplifies general function and avoids extra complexity. In case we decide to go all the way to allocate only a single slot that will need to be changed.

PiperOrigin-RevId: 767785846
Change-Id: Ia66be1044f06ab43d36064edfc8007ea5ae3767e