Pass swisstable seed as seed to absl::Hash so we can save an XOR in H1.
This reverts the optimization in Copy() of not hashing keys for single-group tables because we can no longer depend on H2 being constant for different tables of the same type.

Note that I also tried a version in which we don't change PerTableSeed to use sign-extended loads, but that version had worse performance on loadtests - presumably because the high bits of the seed are all 0. The intuition here is that if the high bits of the seed are all 0, then the high bits of the 128-bit product of the first Mix can also be all 0, which degrades the quality of the state after the first Mix.

PiperOrigin-RevId: 772077751
Change-Id: I5193d90209f4a0c7b41e36a484adb989311aed65