Improve mixing on 32-bit platforms.
The previous version of Mix doesn't have any entropy from the high bits of the input in the low bits of the output. This can be a problem for 64-bit types whose high bits are salient, such as double.

Motivation: when adding debug asserts to detect cases of long probe sequences in absl hash tables, we see assertion failures for the case of inserting [-1024, 1024] into flat_hash_set<double> in 32-bit architectures.
PiperOrigin-RevId: 773741924
Change-Id: I5b718cec08f175376bf46adbec53500956c392cf