Simplify ConvertSpecialToEmptyAndFullToDeleted
Take advantage of the fact that kEmpty is already the smallest sign char value, and using a saturated subtraction will not affect it.

All special values have the MSB set, so after an AND with MSBS, we are left with -128 for special values and 0 for full. After applying subs 2, we arrive at the result of -128(kEmpty) for special and -2(kDeleted) for full.

According to Intel specifications, the subs(PSUBSB) instruction has equal or better CPI and latency than shuffle(PSHUFB), so there is no need to keep the SSSE3 alternative.